Phase 5 commit 1: FreeBsdFileSystemFFM + Sonar fixes
FreeBsdFileSystemFFM mirrors FreeBsdFileSystemJNA — two trivial
overrides for kern.openfiles / kern.maxfiles via BsdSysctlUtilFFM.

Folds in three Sonar findings from PR #3322:
* FreeBsdNetworkParamsFFM (S2184): widen HOST_NAME_MAX + 1 to long for
  the arena allocate call to match the gethostname argument on the
  next line, keeping both arithmetic paths in 64-bit math.
* DisklabelTest x2 (S2153): drop the unnecessary .longValue() unboxing
  from assertThat(diskdata.getC()…) in the OpenBSD and NetBSD live
  smoke tests — Hamcrest's greaterThanOrEqualTo(0L) works on the Long
  directly, the unbox-then-rebox roundtrip is wasted work.

The CPUMASK regex S5998 finding in FreeBsdCentralProcessorFFM (and its
JNA sibling) was reviewed and marked false positive: the comma anchor
between alternatives prevents catastrophic backtracking, and input is
bounded kernel-controlled sysctl output.

+++ oshi-core/src/test/java/oshi/driver/unix/bsd/disk/DisklabelTest.java
@@ -172,7 +172,7 @@ class DisklabelTest {
             if (elevated) {
                 assertThat("Disk label is not null", diskdata.getA(), is(not(nullValue())));
                 assertThat("Disk duid is not null", diskdata.getB(), is(not(nullValue())));
-                assertThat("Disk size is nonnegative", diskdata.getC().longValue(), is(greaterThanOrEqualTo(0L)));
+                assertThat("Disk size is nonnegative", diskdata.getC(), is(greaterThanOrEqualTo(0L)));
                 for (HWPartition part : diskdata.getD()) {
                     assertTrue(part.getIdentification().startsWith(diskName), "Partition ID starts with disk");
                 }
@@ -191,7 +191,7 @@ class DisklabelTest {
             if (elevated) {
                 assertThat("Disk label is not null", diskdata.getA(), is(not(nullValue())));
                 assertThat("Disk duid is not null", diskdata.getB(), is(not(nullValue())));
-                assertThat("Disk size is nonnegative", diskdata.getC().longValue(), is(greaterThanOrEqualTo(0L)));
+                assertThat("Disk size is nonnegative", diskdata.getC(), is(greaterThanOrEqualTo(0L)));
             }
         }
     }
